dataset_api_app/requirements.txt
flask
gunicorn
numpy
pandas
pyarrow
pydantic
python-dotenv
requests
stream-zip
sqlalchemy
werkzeug
